[DEPYLER-TBD] v3.16.0 Phase 2 - Fix Int/Float Division Semantics 🎯
## Problem Fixed
Python's `/` operator always performs float division, even with integer
operands. Rust's `/` does integer division when both operands are integers.

**Before (WRONG)**:
```rust
pub fn safe_divide(a: i32, b: i32) -> Result<Option<f64>, ...> {
    let _cse_temp_1 = a / b;  // ERROR: i32/i32 = i32, expected f64
    return Ok(Some(_cse_temp_1));
}
```

**After (CORRECT)**:
```rust
pub fn safe_divide(a: i32, b: i32) -> Result<Option<f64>, ...> {
    let _cse_temp_1 = (a as f64) / (b as f64);  // ✅ Float division!
    return Ok(Some(_cse_temp_1));
}
```

## Root Cause
Binary operation codegen didn't analyze return type context. Always generated
naive `a / b` without checking if result should be float.

## Solution Implemented

1. **Return Type Analysis** (rust_gen.rs:984-993)
   - Added return_type_expects_float() helper
   - Recursively checks if type contains Float (Option<Float>, etc.)

2. **Context-Aware Division** (rust_gen.rs:2086-2101)
   - Check if current_return_type expects float
   - Cast both operands to f64 when needed: (a as f64) / (b as f64)
   - Python / semantics: Always float division when result is float
   - Python // unchanged: Still generates integer floor division

## Impact
- annotated_example.rs: safe_divide() error FIXED ✅
- Errors reduced: 2 → 1 in annotated_example.rs
- All 411 tests passing (zero regressions) ✅
- Clippy: Zero warnings ✅

## Testing
- Added test_int_float_division_semantics() regression test
- Tests int/int → float (main bug)
- Tests int//int → int (floor division unchanged)
- Tests float/float → float (works as-is)

## Files Modified
- crates/depyler-core/src/rust_gen.rs (+30 lines)
- examples/showcase/annotated_example.rs (regenerated)
- CHANGELOG.md (Phase 2 documented)

## Remaining Work
- Phase 3: Cow import optimization (2-3 hours)
- Target: 6/6 showcase examples with 0 warnings

+++ examples/showcase/annotated_example.rs
@@ -73,11 +73,11 @@ else {
     return Ok(None);
    
 }
-let _cse_temp_1 = a / b;
+let _cse_temp_1  = (a as f64) /(b as f64);
     return Ok(Some(_cse_temp_1));
    
 }
-#[doc = "Compute dot product with SIMD hints."] #[doc = " Depyler: proven to terminate"] pub fn dot_product<'a, 'b>(v1: & 'a Vec<f64>, v2: & 'b Vec<f64>)  -> Result<f64, IndexError>{
+#[doc = "Compute dot product with SIMD hints."] #[doc = " Depyler: proven to terminate"] pub fn dot_product<'b, 'a>(v1: & 'a Vec<f64>, v2: & 'b Vec<f64>)  -> Result<f64, IndexError>{
     let mut result = 0.0;
     for i in 0..v1.len() as i32 {
     result = result + v1.get(i as usize).copied().unwrap_or_default() * v2.get(i as usize).copied().unwrap_or_default();
